% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePvalues.R
\name{calculate_full_set_pvalue}
\alias{calculate_full_set_pvalue}
\title{Calculate Proportion-Based P-Value for Entire Feature Set}
\usage{
calculate_full_set_pvalue(permutedvalues, quantiledata)
}
\arguments{
\item{permutedvalues}{A data frame containing the permuted feature importances.
It should include columns for feature importance and permutation number.
Typically, this data frame is generated by a function that performs permutation tests.}

\item{quantiledata}{A data frame containing quantile information of feature importances.
This data frame should include columns for feature rank, mean permuted importance, observed true importance,
and optionally their logarithmic transformations. This is usually output from \code{calculateQuantiles} function.}
}
\value{
A data frame with a single column "p_val_set".
This column contains the calculated p-value, which quantifies the probability of observing a sum of absolute deviations
as extreme as the one observed, under the null hypothesis. If the p-value is extremely low (below the resolution of the
number of permutations), it is reported as less than the reciprocal of the number of permutations.
}
\description{
This function calculates the proportion-based p-value for the entire feature set by comparing
the observed sum of absolute deviations from the mean feature importance to those from permuted data.
It is particularly useful in permutation tests to assess the statistical significance of the observed data.
}
\examples{
pvalue_set <- calculate_full_set_pvalue(feat_importances$permuted_importances, quantile_data)
}
