% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePvalues.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Calculate Quantiles of Feature Importance}
\usage{
calculate_quantiles(truevalues, permutedvalues, alpha = 0.05)
}
\arguments{
\item{truevalues}{A data frame containing true feature importance data.
It should have columns: feature_rank, feature_importance, and log_feature_importance.
This is typically obtained from the \code{calculate_feature_importances} function. (It is feat_importances$true_importances).}

\item{permutedvalues}{A data frame containing permuted feature importance data.
It should have the same structure as \code{truevalues} and is typically obtained from the \code{calculate_feature_importances} function. (It is feat_importances$permuted_importances).}

\item{alpha}{Numeric value representing the significance level used to calculate the upper and lower quantiles of the permuted importances.
It must be between 0 and 1, exclusive. Default is 0.05.}
}
\value{
A tibble/data frame with columns for feature rank, mean permuted importance, lower and upper quantile bounds,
observed true importance, and their respective logarithmic transformations.
This data frame provides a comprehensive view of where true importances lie in relation to the distribution of permuted importances.
}
\description{
This function computes the mean, lower, and upper quantiles of permuted feature importance scores
and compares them with the observed importance scores from the true values. It is used to
assess the significance of feature importances from random forest models by comparing them
against a distribution of importances obtained through permutation.
}
\examples{
quantile_results <- calculate_quantiles(feat_importances$true_importances, feat_importances$permuted_importances, alpha=0.05)

}
