% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDatasets.R
\name{create_feature_matrix}
\alias{create_feature_matrix}
\title{Dataset Preparation Function for Scikit-Learn Random forest in pyRforest.}
\usage{
create_feature_matrix(dataset, set_type)
}
\arguments{
\item{dataset}{A data frame representing the dataset to be prepared. The dataset must contain the columns 'identifier', and 'target', which will be excluded from the main matrix.}

\item{set_type}{A character string specifying the type of dataset being prepared. Possible values are "training", "testing", or "validation". This parameter influences the naming of the output elements.}
}
\value{
A list containing three elements:
\itemize{
\item{X_set_type}{ A data frame of the dataset after excluding specified columns.}
\item{X_set_type_mat}{ A matrix conversion of 'X_set_type'.}
\item{y_set_type_vector}{ A vector representing the 'target' column from the original dataset.}
}
The names of these elements are dynamically generated based on the 'set_type' parameter.
}
\description{
This function prepares a dataset for scikit-learn random forest machine learning by creating a matrix from the dataset excluding certain columns, and extracting a target vector. It is flexible and can handle different types of data sets such as training, testing, or validation.
Data must be in the correct input format, see vignette or example dataset for details.
}
\examples{
library(pyRforest)

Load sample data
data("demo_rnaseq_data")

processed_training_data <- create_feature_matrix(demo_data_rnaseq_rf$training_data, "training")

processed_validation_data <- create_feature_matrix(demo_data_rnaseq_rf$validation_data, "validation")

processed_testing_data <- create_feature_matrix(demo_data_rnaseq_rf$testing_data, "testing")

}
