% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_rnaseq_data}
\alias{demo_rnaseq_data}
\title{Demo RNA-seq Dataset}
\format{
A list of 4 tibbles
}
\source{
This data is a subset of a larger dataset of real-patient data from study:
SRA Accession: PRJNA788974 mapped to the Hg38 and aligned using STAR Aligner.
Data has been manipulated and is for demonstration purposes only
}
\usage{
demo_rnaseq_data
}
\description{
A dataset containing RNA-seq data for demonstration purposes.
}
\keyword{datasets}
