% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigures.R
\name{generate_fi_rank_plot}
\alias{generate_fi_rank_plot}
\title{Generate Feature Importance Score-Rank Plot}
\usage{
generate_fi_rank_plot(
  permutedvalues,
  quantiledata,
  xlimitmin = 1,
  xlimitmax = 500,
  ylimitmin = -10,
  ylimitmax = 0,
  xtickbreaks = 10,
  labelverticaladjust = 1.05,
  labelhorizontaladjust = 1.05,
  indvPermScoresOn = TRUE,
  logOn = TRUE,
  focusedView = FALSE
)
}
\arguments{
\item{permutedvalues}{a tibble of feature ranks, feature names, feature importance scores, log feature importance scores and permutation # generated by the load_permuted_fws function}

\item{quantiledata}{a tibble containing: feature ranks, mean, lower, upper, and observed as well as logmean, loglower, logupper, logobserved generated by the calculateQuantiles function}

\item{xlimitmin}{a numerical value for the minimum x value in the figure, default is 1 because ranking starts at 1.}

\item{xlimitmax}{a numerical value for the maximum x value in the figure, default is 500 (this is effectively the total number of features to include)}

\item{ylimitmin}{a numerical value for the minimum y value in the figure, default is -10}

\item{ylimitmax}{a numerical value for the maximum y value in the figure, default is 0, but this will vary largely depending on your dataset, set it larger then narrow down}

\item{xtickbreaks}{a numerical value for the break spacing between label ticks on the x axis, default is 10}

\item{labelverticaladjust}{a numerical value specifying the vertical (y-axis) adjustment of the intersection with upper quartile label}

\item{labelhorizontaladjust}{a numerical value specifying the horizontal (x-axis) adjustment of the intersection with upper quartile label}

\item{indvPermScoresOn}{logical TRUE or FALSE, whethere or not to include the lines for each individual permutation feature importance score (shown in light blue)}

\item{logOn}{logical TRUE or FALSE, whether or not to use the log or unlogged data. Default to TRUE.}

\item{focusedView}{logical TRUE or FALSE, whether or not to show the full plot or a focused view of only the features above the cutoff line.}
}
\value{
A plot of log feature importance score by feature rank showing the upper quartile intersection which we recommend to use for a cutoff.
}
\description{
This function generates a figure which shows the true observed data plotted against the permuted data, by rank. The intersection of the true data with the upper quartile is shown, which we recommend as a significance cutoff.
Note: There are ample parameters for controlling the axes scale, label location, and zoom, because of data variability, you will almost certainly have to adjust these to fit your plot.
}
\examples{
fiplot_full <- generate_fi_rank_plot(feat_importances$permuted_importances, quantile_data, xlimitmin=1, xlimitmax = 10, ylimitmin= -5, ylimitmax = 0, labelhorizontaladjust = -0.05,labelverticaladjust = 1.5,focusedView = FALSE,logOn = TRUE)
fiplot_focused <- generate_fi_rank_plot(feat_importances$permuted_importances, quantile_data, xlimitmin=1, xlimitmax = 10, ylimitmin= -5, ylimitmax = 0, labelhorizontaladjust = 1.05,labelverticaladjust = 1.5,focusedView = TRUE,logOn = TRUE)
}
