% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigures.R
\name{generate_pi_ECDF_plot}
\alias{generate_pi_ECDF_plot}
\title{Generate ECDF Plot}
\usage{
generate_pi_ECDF_plot(permutedvalues, quantiledata)
}
\arguments{
\item{permutedvalues}{a tibble of feature ranks, feature names, feature importance scores, log feature importance scores and permutation # generated by the load_permuted_fws function}

\item{quantiledata}{a tibble containing: feature ranks, mean, lower, upper, and observed as well as logmean, loglower, logupper, logobserved generated by the calculateQuantiles function}
}
\value{
A plot of the empirical cumulative density function showing the sum of absolute deviations by fraction of data.
}
\description{
Creates a plot of the empirical cumulative density function showing the sum of absolute deviations by fraction of data.
}
\examples{
ecdf_plot<-generate_pi_ECDF_plot(feat_importances$permuted_importances, quantile_data)
}
