% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigures.R
\name{generate_pi_histogram}
\alias{generate_pi_histogram}
\title{Generate pi Histogram Plot}
\usage{
generate_pi_histogram(permutedvalues, quantiledata)
}
\arguments{
\item{permutedvalues}{a tibble of feature ranks, feature names, feature importance scores, log feature importance scores and permutation # generated by the load_permuted_fws function}

\item{quantiledata}{a tibble containing: feature ranks, mean, lower, upper, and observed as well as logmean, loglower, logupper, logobserved generated by the calculateQuantiles function}
}
\value{
A histogram of count by sum of absolute deviations, showing the differences between the null and observed datasets.
}
\description{
Creates a histogram showing the sum of absolute deviations by count, the measures used to calculate the pi statistics used in the p-value calculation for the set.
}
\details{
This function generates a histogram illustrating the sum of absolute deviations in the permuted set vs the true (observed) set. It is a visualization of the data used to calculate the p-value for the entire feature set.
}
\examples{
pihist_plot<-generate_pi_histogram(feat_importances$permuted_importances, quantile_data)
}
