% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigures.R
\name{generate_shap_plots}
\alias{generate_shap_plots}
\title{Generate SHAP Plots}
\usage{
generate_shap_plots(
  mean_shap_values,
  long_shap_data,
  title_global = "SHAP Global Feature Importance",
  title_local = "SHAP Local Feature Explanation",
  fill_colors = c("blue", "red"),
  gradient_colors = c("blue", "red")
)
}
\arguments{
\item{mean_shap_values}{A dataframe containing mean SHAP values and features}

\item{long_shap_data}{A dataframe containing individual SHAP values for features}

\item{title_global}{The title for the global feature importance plot}

\item{title_local}{The title for the local feature explanation plot}

\item{fill_colors}{A vector of colors to use for filling the bar plot}

\item{gradient_colors}{A vector of colors to use for the color gradient in the beeswarm plot}
}
\value{
A combined ggplot object with a bar plot and a beeswarm plot
}
\description{
Creates a combined bar and beeswarm plot to show global and local feature importance
}
\examples{
shapplot<-generate_shap_plots(mean_shap_values = shapvals$significant_features,long_shap_data = shapvals$long_shap_data)
}
